pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "cv-agent"
version = "0.1.0"
description = "AI-powered CV and cover letter generation system"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "CV Agent Team"}
]
dependencies = [
    # Core dependencies
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "ruamel.yaml>=0.18.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",

    # Graph visualization
    "graphviz>=0.21",

    # Translation services
    "googletrans>=4.0.0rc1",

    # LangChain and RAG (optional but recommended)
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "langchain-openai>=0.0.5",
    "chromadb>=0.4.0",

    # PDF generation for cover letters
    "reportlab>=4.0.0",

    # OpenAI for LLM features (optional)
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "ruff",
]

[project.scripts]
cv-agent = "main:app"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "COM812", "ISC001"]

[tool.black]
line-length = 88
target-version = ['py39']
